refactored src folder structure to contain modules and root script(s) only
