--- components/server/src/__init__.py
@@ -1,3 +1,0 @@
-"""Import entry point."""
-
-from .serve import serve
