+++ components/server/src/database/datamodels.py
@@ -1,11 +1,10 @@
 """Datamodels collection."""
 
-from typing import Any, Dict
-
 import pymongo
-from pymongo.database import Database
 
-from ..util import iso_timestamp
+from pymongo.database import Database
+from typing import Any, Dict
+from utilities.functions import iso_timestamp
 
 
 def latest_datamodel(database: Database, max_iso_timestamp: str = ""):
