+++ components/server/src/database/measurements.py
@@ -1,12 +1,12 @@
 """Measurements collection."""
 
-from datetime import date
-from typing import Optional
-
 import pymongo
-from pymongo.database import Database
 
-from ..util import iso_timestamp
+from datetime import date
+from pymongo.database import Database
+from typing import Optional
+from utilities.functions import iso_timestamp
+
 from .datamodels import latest_datamodel
 from .reports import latest_metric
