+++ components/server/src/initialization/__init__.py
@@ -1,4 +1,4 @@
 """Initialization functions for bottle, the database, and LDAP."""
 
-from .database import init_database
 from .bottle import init_bottle
+from .database import init_database
