+++ components/server/src/routes/auth.py
@@ -1,18 +1,18 @@
 """Login/logout."""
 
-from datetime import datetime, timedelta
-from typing import Dict
+import bottle
+import ldap  # pylint: disable=import-error,wrong-import-order
 import logging
 import os
 import re
 import urllib.parse
 
+from datetime import datetime, timedelta
+from typing import Dict
 from pymongo.database import Database
-import bottle
-import ldap  # pylint: disable=import-error,wrong-import-order
+from utilities.functions import uuid
 
 from ..database import sessions
-from ..util import uuid
 
 
 def generate_session_id() -> str:
