+++ components/server/src/routes/cors.py
