+++ components/server/src/routes/plugins/__init__.py
@@ -1,0 +1,2 @@
+from .authentication_plugin import AuthenticationPlugin
+from .injection_plugin import InjectionPlugin
