+++ components/server/src/routes/plugins/authentication_plugin.py
@@ -1,10 +1,9 @@
 """Route authentication plugin."""
 
+import bottle
 import logging
 
-import bottle
-
-from ..database import sessions
+from database import sessions
 
 
 class AuthenticationPlugin:  # pylint: disable=too-few-public-methods
